[DRAFT][NO MERGE] Refactoring VerifyQuantizedNode*Type
This draft if for refactoring `VerifyQUantizedNote*Type`.

ONE-DCO-1.0-Signed-off-by: Seok NamKoong <seok9311@naver.com>
